Avoid sending a pokemon's clone hash to the client
Previously, the clone hash was considered 'private data' along with the PID (since if the clone hash was known, the PID could be brute-forced), so it was only sent if the client was the owner of the pokemon, or if the pokemon was public.

However, there's not really any reason to send the clone hash to the client at all, so it's better to just mark it as a private property and not ever send it to the client.