Do not send a PID hint
If the server sends the first half of the PID to everyone, and it also sends the ESV, then there are only 16 possible values for the PID. This makes it possible to determine the PID by uploading a .pk6 file for each of the 16 possible PIDs until it registers as a clone.

To avoid the issue, the server should simply not send a PID hint. The benefit of the PID hint (i.e. determining whether two things have roughly the same PID) can be replicated by checking the ESV anyway.