Add API support for gen 7 (#618)
* Use _rawFile instead of _rawPk6 for raw files

* Add API support for gen 7

This allows a `gen` property to be specified from the multi-upload API, defaulting to 6. The property is passed as an option to pk6parse (which currently doesn't do anything with it), and it's sent as a regular property to the client.

Uploads are only allowed to have a gen that is in the `pk6parse.SUPPORTED_GENS` array, or `[6]` is `pk6parse.SUPPORTED_GENS` doesn't exist (which is the case for the old version of pk6parse. This means that gen 7 uploads are currently not allowed (since someone could upload a file with `gen: 7` that fails future validation checks). To enable them, we should export `SUPPORTED_GENS = [6, 7]` from pk6parse when the parser is ready, which will make the server accept gen 7 uploads.